correção de erros , feito teste

+++ src/main/java/com/example/gabrieljony/projetoCadastroapi/repository/ClienteRepository.java
@@ -1,6 +1,11 @@
 package com.example.gabrieljony.projetoCadastroapi.repository;
 
+import com.example.gabrieljony.projetoCadastroapi.model.Cliente;
+import org.springframework.data.repository.CrudRepository;
+
+import java.util.List;
+
 public interface ClienteRepository extends CrudRepository<Cliente, Long> {
-    List<Cliente> findByCpf(int cpf);
+    List<Cliente> findByCpf(long cpf);
 
 }
